Create split-jar-mergeJson.ps1
this ps1 split the item which is a big string and this string has strings which are separated by space. such as file paths. 
each splitted part is executed by jar and the result is collected on json file. At last, json files are merged. 
I developed this code for azure devops build pipeline that's why, it has $env variables.